SOF-142: Refactor all complete session files
