+++ app/src/main/java/com/vci/vectorcamapp/complete_session/list/presentation/CompleteSessionState.kt
@@ -1,4 +1,4 @@
-package com.vci.vectorcamapp.complete_session.presentation
+package com.vci.vectorcamapp.complete_session.list.presentation
 
 import com.vci.vectorcamapp.core.domain.model.Session
